requirements.txt
annotated-types==0.6.0
attrs==23.1.0
audio2numpy==0.1.2
audioread==3.0.1
Babel==2.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clldutils==3.20.0
colorama==0.4.6
colorlog==6.7.0
csvw==3.2.1
decorator==5.1.1
dlinfo==1.2.1
einops==0.7.0
ffmpeg==1.4
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.4
inflect==7.0.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
language-tags==1.2.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.7
munch==4.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
packaging==23.2
phonemizer==3.2.1
Pillow==10.1.0
platformdirs==4.0.0
pooch==1.8.0
progressbar==2.5
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pylatexenc==2.10
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
rotary-embedding-torch==0.3.5
rpds-py==0.13.1
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
segments==2.2.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.12.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tortoise-tts==3.0.0
tqdm==4.66.1
transformers==4.19.0
triton==2.1.0
typing_extensions==4.8.0
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.1.0
vector-quantize-pytorch==1.11.7
